Ensure correct objective data is pulled for each map
Before the Desert Borderlands were added each of the BLs had it's own set of IDs from the API. Now with the Desert Borderlands there is one set of IDs that is the same across the BLs. This caused some of the match updating code to break because it was looking for unique objectives for each BL. To solve this I added a check to ensure that the map for the objective we got from the API and the map for the objective we are updating are the same.