Add support for 64bit client and cache the Gw2Process value
The new method of getting the GW2 process searches all process for ones that start with "Gw2" so it cataches both the 32bit and 64bit clients. This method is a little more intensive so it only gets the process when it doesn't have one or the one it had has exited.